+++ examples/mailboxNoRouter/index.js
@@ -1,9 +1,9 @@
-require('../styles.css')
-const F = require('../../lib')
-const io = require('socket.io-client')
+import '../styles.css'
+import F from '../../lib'
+import io from 'socket.io-client'
 
-const realData = require('./data.service') // real data service
-const mockedData = require('./mocked-data.service') // mocked data service
+import realData from './data.service' // real data service
+import mockedData from './mocked-data.service' // mocked data service
 
 // realData | mockedData, for a real server uncomment socketio
 const data = realData
@@ -12,7 +12,8 @@ let socket = io(data.serverName, {})
 data.connect(socket)
 
 let engine = F.run({
-  root: F.log(require('./mailbox')),
+  logAll: true,
+  root: require('./mailbox').default,
   services: {
     data,
   },
@@ -30,7 +31,7 @@ if (module.hot) {
   // We accept updates to the top component
   module.hot.accept('./mailbox', (comp) => {
     // Mutate the variable holding our component
-    let module = require('./mailbox')
+    let module = require('./mailbox').default
     engine.reattach(module)
   })
 }