+++ examples/mailboxNoRouter/mocked-data.service.js
@@ -2,7 +2,7 @@
 /*
   this module contains all data fetching and inyecting logic
 */
-const F = require('../../lib')
+import F from '../../lib'
 
 let serverName = 'http://localhost:4000'
 let pageSize = 50
@@ -11,7 +11,7 @@ let messages = getInitialMessages()
 
 let messagesReceivedCb = () => 0
 
-module.exports = F.service({
+export default F.service({
   serverName,
   queue: true,
   store: {